requirements.txt
fastapi[0.68.1]
uvicorn[standard]
mysql-connector[standard]
argon2-cffi[standard]
